Produces BayesTraits linked input data

It produces all BayesTraits input data of linked samples of trait/climate data (distribution data) for phylogenetic regression analyses with Meade & Pagel's (2022) BayesTraits computer package. The function will automatically generate two .txt files: one including the linked traits and the other with mean values for each included trait. The function also includes an argument to calculate the net node count for each species/accession given a phylogenetic tree of reference, which has been successfully used to explore the effect of speciation rates (e.g. O’Donovan et al. 2018) on the response variable. See BayesTraits V4.0.0 Manual for further details on the specific format of the input files that are required by BayesTraits.

Usage

BayesTraits.inputs(tree,
                   data,
                   tipscol = NULL,
                   NodeCount = FALSE,
                   logtransf = NULL,
                   sqrtransf = NULL,
                   traitcols = NULL,
                   addtraits = NULL,
                   ordtraits = NULL,
                   dir_create = "results_BayesTraits_input",
                   fileDistData = "BayesTraits_linked_input.txt",
                   fileMeanData = "BayesTraits_mean_input.txt",
                   fileOrigData = NULL)

Arguments

tree

The input tree file. This can be either a Newick-formatted tree imported by ape's function read.tree or a treedata-formatted phylogeny as generated by BEAST and read by treeio's function read.beast.

data

The input sample of trait (distribution) data in data.frame format. A column must be included that links each accession/record and its corresponding sample of trait data with a terminal branch label of the input phylogeny.

tipscol

Give the name of the column with terminals/tip labels within the input distribution of trait data.

NodeCount

Logical, the default is FALSE. The function will create a new column NodeCount in the input trait data and place it just to the left of the column longitude. The net node count for each species (replicated for each accession within the same species) derives from the input tree.

logtransf

A vector with the name of the column(s) of the trait data to be log10-transformed. For example, if you define the columns c("DBH", "NodeCount") to be log10-transformed, then the function will automatically create the new columns log10DBH and logNodeCount from the log10-transformed DBH and NodeCount columns in the original input data.

sqrtransf

A vector with the name of the column(s) of trait data to be square log10-transformed. For example, if you define the column Nodecount to be transformed, then the function will automatically create the new column sqrNodeCount from the NodeCount column in the original input data. If this trait is to be included in an analysis, list in as argument in addtraits (see below).

traitcols

A vector with the name of the column(s) of the input sample of trait data that will serve as the response/dependent variable (first column name) and the explanatory/independent variable(s) (second and possibly a third and fourth column) for a phylogenetic regression analysis in BayesTraits.

addtraits

A vector with the name of additional column(s) that are not originally in the input trait data but for which you want to generate inputs of linked samples of trait data and mean data for the BayesTraits analysis. For example, if you have chosen NodeCount=TRUE and defined any new columns to be log10-transformed or square-transformed, they should be listed here as additional columns to be included in the linked sample of trait data.

ordtraits

A vector listing the order of trait names to be used in the BayesTraits analysis. This ensures that the first trait column will serve as the response variable, and that subsequent trait columns serve as explanatory variables. Maintain a consistent order of names associated with explanatory variables to interpret the BayesTraits log output file where Beta1, Beta2, etc. are reported. Coefficient estimates reported under Beta1 can be associated with the first explanatory variable listed, and so on.

dir_create

Pathway to the computer's directory where the file will be saved. The default setting creates a directory named results_BayesTraits_input and the results will be saved within a subfolder of that directory named with the current date.

fileDistData

Name of the resulting .txt file of linked samples of trait data. If no name is given, the default setting creates a file named BayesTraits_linked_input.txt.

fileMeanData

Name of the resulting .txt file of mean trait data. If no name is given, the default setting creates a file named BayesTraits_mean_input.txt.

fileOrigData

If you define any name here, the originally imported trait data will be saved as a CSV file, including any newly generated data columns. The trait data will be written on disk within the directory as set in the argument dir_create, but outside the current date subfolder.

See also

Author

Domingos Cardoso & Matt Lavin

Examples

if (FALSE) {
library(InNOutBT)

## Loading a example data file
data(vatsData)

## Loading a example tree file
data(vatsTree)

BayesTraits.inputs(tree = vatsTree,
                   data = vatsData,
                   tipscol = "terminal",
                   NodeCount = TRUE,
                   logtransf = c("DBH", "NodeCount"),
                   traitcols = c("bio12", "bio15"),
                   addtraits = c("log10NodeCount", "log10DBH"),
                   ordtraits = c("log10DBH", "bio12", "bio15", "log10NodeCount"),
                   dir_create = "results_BayesTraits_input",
                   fileDistData = "BayesTraits_linked_data_bio12_bio15_nnodes.txt",
                   fileMeanData = "BayesTraits_mean_data_bio12_bio15_nnodes.txt",
                   fileOrigData = "vataireoids_1610_25May2022_BayesTraits_netnodes_logtransf.csv")
}