Process BayesTraits output data

It processes the output Log and Stones .txt files from phylogenetic regression analyses with Meade & Pagel's (2022) BayesTraits computer package. The resulting processed files are tables either in Word or CSV formats summarizing each model, including comparison with Bayes Factor. Heatmaps based on pairwise Bayes Factor comparison are also automatically produced in PDF format. See BayesTraits V4.0.0 Manual for further details on each of the output files produced by BayesTraits so as to better understand the processed output files from the current function BayesTraits.outputs.

Usage

BayesTraits.outputs(logst_dir = NULL,
                    responvar = NULL,
                    explanvar = NULL,
                    treetransf = c("Kappa", "Delta", "Lambda", "OU"),
                    bayesfactor = FALSE,
                    unirates = TRUE,
                    outformat = c("Word", "CSV"),
                    tableleg = "Phylogenetic regression in BayesTraits: models and coefficients.",
                    dir_create = "results_BayesTraits_output",
                    outfile = "BayesTraits_output_table",
                    ...)

Arguments

logst_dir

Path to the folder directory where the Log and stepping stone sampler files from the BayesTraits phylogenetic independent contrast regression analyses are stored. The Log file(s) contain the model options and output. The stepping stone sampler files file(s) contain the estimated log marginal likelihood and the additional information about the stones. Make sure that all these are also named with the explanatory variable(s) separated by underscore, as well as the exact name of the tree transformation set in the analysis. So, you have to name them either with Kappa, Delta, Lambda, OU (Ornstein–Uhlenbeck), UNI (if you have used an uniform tree transformation), VR (if you have used a variable rates) or Fabric (if you have used the general statistical model that accommodates uneven evolutionary landscape as described by Pagel et al. (2022). Thus, your Log file(s) should be name as something like _VAR1NAME_VAR2NAME.txt.Log.TREETRANSFORMATIONNAME.txt. See the following real examples of Log input files: "BayesTraits_mean_data_bio12_bio15.txt.Log.VR" and "BayesTraits_mean_data_bio12_bio15_nnodes.txt.Log.Lambda". This means that the first Log file contains two models named "bio12" and "bio15" and was based on VR (Variable Rates tree transformation), the second Log file contains three models named "bio12", "bio15", and "nnodes", and was based on Lambda tree transformation. Likewise, your Stones file(s) should be name as something like _VAR1NAME_VAR2NAME.txt.Stones.TREETRANSFORMATIONNAME.txt. See the following real examples of Stones input files: "BayesTraits_mean_data_bio12_bio15.txt.Stones.VR" and "BayesTraits_mean_data_bio12_bio15_nnodes.txt.Stones.Lambda". This means that the first Stones file contains two models named "bio12" and "bio15" and was based on VR (Variable Rates tree transformation), the second Stones file contains three models named "bio12", "bio15", and "nnodes", and was based on Lambda tree transformation.

responvar

Give the name of the response variable.

explanvar

Give the name(s) of the explanatory variable(s) exactly as they are also written in the names of the Log and Stones files to be processed.

treetransf

Select any of the available BayesTraits tree transformation of evolution (Kappa, Delta, Lambda or OU), with which you have analysed your data. If you have analysed your data with just Uniform, VR or Fabric models then make sure you defined treetransf = NULL.

bayesfactor

Logical, if FALSE, then no Bayes Factor pairwise comparisons will performed among the models.

unirates

Logical, FALSE, then the results with uniform tree transformation will not be displayed in the main table.

outformat

Define either "Word" or "CSV", or a vector c("Word", "CSV") of both, for writing the results in any of such formats.

tableleg

A legend for the main table, provided that you have also chosen "Word" in the argument outformat for the resulting format. If you do not give any specific legend here, then the Word-formatted table will have the default legend Phylogenetic regression in BayesTraits: models and coefficients.

dir_create

Pathway to the computer's directory, where the file(s) will be saved. The default setting creates a directory named results_BayesTraits_output in which the results will be saved within another subfolder named by the current date.

outfile

Name of the main resulting table-formatted files in either "Word" or "CSV" (or both formats; i.e. depending on the specified argument outformat) to be saved on the directory. If no name is given, the default setting creates a file named BayesTraits_output_table.docx and/or BayesTraits_output_table.csv.

...

additional parameters passed to pdf.

Value

Table in dataframe which is also saved on disk in .csv format.

See also

Author

Domingos Cardoso & Matt Lavin