Create a BayesTraits project folder with associated example scripts and input data for a full shell script run and log processing of BayesTraits phylogenetic regression analyses

By reporting a directory path and the computer's operating system ("unix" or "windows"), this function creates a BayesTraits project folder with associated example scripts and input data for a full shell script run and log processing multiple phylogenetic regression analyses with Meade & Pagel's (2022) BayesTraits program. The function produces an example of a shell script in Unix or Windows PowerShell that helps to fully automatize both the complete BayesTraits regression analysis and processing of the resulting log and stones files, by automatically running the three main InNOutBT functions phyreg.inputs, phyreg.shell, and phyreg.outputs, without directly opening the R or R Studio programs. Note that these files are just examples from an specific study of the Vataireoid legumes, and so you will have to edit the folder paths and analysis parameters in such example scripts according to your desired regression analyses. See also a more complete article describing a BayesTraits shell run project that can be created with shell.phyreg.proj for fully automated multiple phylogenetic regression analyses.

Usage

shell.phyreg.proj(dir_create_proj = NULL,
                  syst = NULL)

Arguments

dir_create_proj

Path to any computer's directory where an example of a BayesTraits shell project folder and associated files for a full run with shell script should be saved.

syst

Report the operating system ("unix" or "windows").

See also

Author

Domingos Cardoso & Matt Lavin

Examples

if (FALSE) {
library(InNOutBT)

shell.phyreg.proj(dir_create_proj = "/Users/domingoscardoso",
                  syst = "unix")
}